refactor: standardize type imports and restructure GraphQL context and schema handling
